+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: hsilverb <hsilverb@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/16 18:24:31 by hsilverb          #+#    #+#             */
-/*   Updated: 2022/11/25 16:26:02 by hsilverb         ###   ########lyon.fr   */
+/*   Updated: 2022/11/30 16:10:08 by hsilverb         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	unsigned long	i;
-	char			*char_dest;
-	char			*char_src;
+	size_t	i;
+	char	*char_dest;
+	char	*char_src;
 
 	char_dest = (char *)dest;
 	char_src = (char *)src;
