+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: hsilverb <hsilverb@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/15 16:23:11 by hsilverb          #+#    #+#             */
-/*   Updated: 2022/11/19 17:43:30 by hsilverb         ###   ########lyon.fr   */
+/*   Updated: 2022/11/30 16:10:57 by hsilverb         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	unsigned long int	i;
-	char				*str;
+	size_t	i;
+	char	*str;
 
 	i = 0;
 	str = b;
